+++ vamp-plugins/dsp/Filter.h
@@ -22,8 +22,8 @@
 
 struct FilterConfig{
     unsigned int ord;
-    double* ACoeffs;
-    double* BCoeffs;
+    float* ACoeffs;
+    float* BCoeffs;
 };
 
 class Filter  
@@ -34,7 +34,7 @@ public:
 
     void reset();
 
-    void process( double *src, double *dst, unsigned int length );
+    void process( float *src, float *dst, unsigned int length );
 	
 
 private:
@@ -43,11 +43,11 @@ private:
 
     unsigned int m_ord;
 
-    double* m_inBuffer;
-    double* m_outBuffer;
+    float* m_inBuffer;
+    float* m_outBuffer;
 
-    double* m_ACoeffs;
-    double* m_BCoeffs;
+    float* m_ACoeffs;
+    float* m_BCoeffs;
 };
 
 #endif
