+++ vamp-plugins/dsp/GetKeyMode.cpp
@@ -15,13 +15,13 @@
 #include <cstdlib>
 
 // Chords profile
-static double MajProfile[36] = 
+static float MajProfile[36] = 
 { 0.0384, 0.0629, 0.0258, 0.0121, 0.0146, 0.0106, 0.0364, 0.0610, 0.0267,
   0.0126, 0.0121, 0.0086, 0.0364, 0.0623, 0.0279, 0.0275, 0.0414, 0.0186, 
   0.0173, 0.0248, 0.0145, 0.0364, 0.0631, 0.0262, 0.0129, 0.0150, 0.0098,
   0.0312, 0.0521, 0.0235, 0.0129, 0.0142, 0.0095, 0.0289, 0.0478, 0.0239};
 
-static double MinProfile[36] =
+static float MinProfile[36] =
 { 0.0375, 0.0682, 0.0299, 0.0119, 0.0138, 0.0093, 0.0296, 0.0543, 0.0257,
   0.0292, 0.0519, 0.0246, 0.0159, 0.0234, 0.0135, 0.0291, 0.0544, 0.0248,
   0.0137, 0.0176, 0.0104, 0.0352, 0.0670, 0.0302, 0.0222, 0.0349, 0.0164,
@@ -33,7 +33,7 @@ static double MinProfile[36] =
 // Construction/Destruction
 //////////////////////////////////////////////////////////////////////
 
-GetKeyMode::GetKeyMode( int sampleRate, float tuningFrequency, double hpcpAverage, double medianAverage ) :
+GetKeyMode::GetKeyMode( int sampleRate, float tuningFrequency, float hpcpAverage, float medianAverage ) :
     m_hpcpAverage( hpcpAverage ),
     m_medianAverage( medianAverage ),
     m_ChrPointer(0),
@@ -51,7 +51,7 @@ GetKeyMode::GetKeyMode( int sampleRate, float tuningFrequency, double hpcpAverag
         
     // Chromagram configuration parameters
     m_ChromaConfig.normalise = MathUtilities::NormaliseUnitMax;
-    m_ChromaConfig.FS        = lrint(sampleRate/(double)m_DecimationFactor);
+    m_ChromaConfig.FS        = lrint(sampleRate/(float)m_DecimationFactor);
     if (m_ChromaConfig.FS < 1) m_ChromaConfig.FS = 1;
 
     // Set C (= MIDI #12) as our base :
@@ -83,16 +83,16 @@ GetKeyMode::GetKeyMode( int sampleRate, float tuningFrequency, double hpcpAverag
     m_MedianBufferFilling = 0;
 
     // Spawn objectc/arrays
-    m_DecimatedBuffer = new double[m_ChromaFrameSize];
+    m_DecimatedBuffer = new float[m_ChromaFrameSize];
     
-    m_ChromaBuffer = new double[m_BPO * m_ChromaBuffersize];
+    m_ChromaBuffer = new float[m_BPO * m_ChromaBuffersize];
     std::fill ( &m_ChromaBuffer[0],&m_ChromaBuffer[m_BPO * m_ChromaBuffersize], 0 );
     
-    m_MeanHPCP = new double[m_BPO];
+    m_MeanHPCP = new float[m_BPO];
     
-    m_MajCorr = new double[m_BPO];
-    m_MinCorr = new double[m_BPO];
-    m_Keys  = new double[2*m_BPO];
+    m_MajCorr = new float[m_BPO];
+    m_MinCorr = new float[m_BPO];
+    m_Keys  = new float[2*m_BPO];
     
     m_MedianFilterBuffer = new int[ m_MedianWinsize ];
     std::fill ( &m_MedianFilterBuffer[0],&m_MedianFilterBuffer[m_MedianWinsize],0 );
@@ -103,7 +103,7 @@ GetKeyMode::GetKeyMode( int sampleRate, float tuningFrequency, double hpcpAverag
     m_Decimator = std::make_unique< Decimator>
         ( m_ChromaFrameSize*m_DecimationFactor, m_DecimationFactor );
 
-    m_keyStrengths = new double[24];
+    m_keyStrengths = new float[24];
 }
 
 GetKeyMode::~GetKeyMode()
@@ -122,17 +122,17 @@ GetKeyMode::~GetKeyMode()
     delete[] m_keyStrengths;
 }
 
-double GetKeyMode::krumCorr(double *pData1, double *pData2, size_t length)
+float GetKeyMode::krumCorr(float *pData1, float *pData2, size_t length)
 {
-    double retVal= 0.0;
+    float retVal= 0.0;
     
-    double num = 0;
-    double den = 0;
-    double mX = MathUtilities::mean( pData1, length );
-    double mY = MathUtilities::mean( pData2, length );
+    float num = 0;
+    float den = 0;
+    float mX = MathUtilities::mean( pData1, length );
+    float mY = MathUtilities::mean( pData2, length );
     
-    double sum1 = 0;
-    double sum2 = 0;
+    float sum1 = 0;
+    float sum2 = 0;
     
     for( size_t i = 0; i <length; i++ )
     {
@@ -153,7 +153,7 @@ double GetKeyMode::krumCorr(double *pData1, double *pData2, size_t length)
     return retVal;
 }
 
-int GetKeyMode::process(double *PCMData)
+int GetKeyMode::process(float *PCMData)
 {
     int key;
 
@@ -193,7 +193,7 @@ int GetKeyMode::process(double *PCMData)
     //calculate mean 		
     for( k = 0; k < m_BPO; k++ )
     {
-        double mnVal = 0.0;
+        float mnVal = 0.0;
         for( j = 0; j < m_ChromaBufferFilling; j++ )
         {
             mnVal += m_ChromaBuffer[ k + (j*m_BPO) ];
@@ -247,9 +247,9 @@ int GetKeyMode::process(double *PCMData)
   }
   std::cout << std::endl;
 */
-    double dummy;
+    float dummy;
     // '1 +' because we number keys 1-24, not 0-23.
-    key = 1 + (int)ceil( (double)MathUtilities::getMax( m_Keys, 2* m_BPO, &dummy )/3 );
+    key = 1 + (int)ceil( (float)MathUtilities::getMax( m_Keys, 2* m_BPO, &dummy )/3 );
 //    std::cout << "key pre-sorting: " << key << std::endl;
     //Median filtering
     // track Median buffer initial filling
@@ -278,7 +278,7 @@ int GetKeyMode::process(double *PCMData)
   std::cout << std::endl;
 */
     auto sortlength = m_MedianBufferFilling;
-    auto midpoint = static_cast<off_t>(ceil(static_cast<double>(sortlength)/2));
+    auto midpoint = static_cast<off_t>(ceil(static_cast<float>(sortlength)/2));
 //  std::cout << "midpoint = " << midpoint << endl;
     if( midpoint <= 0 ) midpoint = 1;
     key = m_SortedBuffer[midpoint-1];
